+++ webview/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@quantumblack/kedro-viz": "^9.2.0",
+    "my-kviz": "^0.0.1-alpha.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1"
